Read excel and csv files
Script to load csv file and excel file and to combine all of the worksheets within the EDM 2022 .xlsx file into one dataframe. Next steps are to change name of column 'EA Permit Reference (EA Consents Database)' in the xlsx file to 'PERMIT_NUMBER' and join dataframes based on this.